+++ src/app/mypage/myprofile/page.tsx
@@ -1,7 +1,7 @@
 'use client';
 import React, { useState } from 'react';
-import UserInfoForm from '../../components/UserInfoForm';
-import SideMenu from '../../components/SideMenu';
+import UserInfoForm from '../../components/mypage/UserInfoForm';
+import SideMenu from '../../components/mypage/SideMenu';
 import AlertModal from '../../components/modal/AlertModal';
 import ConfirmModal from '../../components/modal/ConfirmModal';
 import styles from '../mypage.module.scss';
